Examples: move particle selector to truth tracking module
