+++ Examples/Algorithms/Generators/CMakeLists.txt
@@ -2,7 +2,6 @@ add_library(
   ActsExamplesGenerators SHARED
   ActsExamples/Generators/EventGenerator.cpp
   ActsExamples/Generators/ParametricParticleGenerator.cpp
-  ActsExamples/Generators/ParticleSelector.cpp
   ActsExamples/Options/ParticleGunOptions.cpp)
 target_include_directories(
   ActsExamplesGenerators
