+++ Examples/Algorithms/TruthTracking/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(
   ActsExamplesTruthTracking SHARED
+  ActsExamples/TruthTracking/ParticleSelector.cpp
   ActsExamples/TruthTracking/ParticleSmearing.cpp
   ActsExamples/TruthTracking/TrackSelector.cpp
   ActsExamples/TruthTracking/TruthTrackFinder.cpp
