+++ Examples/Run/Vertexing/IterativeVertexFinderExample.cpp
@@ -9,12 +9,12 @@
 #include "Acts/EventData/TrackParameters.hpp"
 #include "ActsExamples/Framework/Sequencer.hpp"
 #include "ActsExamples/Generators/FlattenEvent.hpp"
-#include "ActsExamples/Generators/ParticleSelector.hpp"
 #include "ActsExamples/Generators/Pythia8ProcessGenerator.hpp"
 #include "ActsExamples/Io/Csv/CsvParticleWriter.hpp"
 #include "ActsExamples/Io/Root/RootParticleWriter.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Options/Pythia8Options.hpp"
+#include "ActsExamples/TruthTracking/ParticleSelector.hpp"
 #include "ActsExamples/TruthTracking/TrackSelector.hpp"
 #include "ActsExamples/TruthTracking/TruthVerticesToTracks.hpp"
 #include "ActsExamples/Utilities/Paths.hpp"
